fix: cast ICA variance to double before storing in pipeline report
EEG.data from some recording systems is stored as single precision.
var() and TESA's compVars inherit that type, so variance percentages
written into fileReport.ica.varRemoved were single. Meanwhile
initPipelineReport initialises varRemoved as NaN (double). When
summarizeReports ran cellfun(@(r) r.ica.varRemoved, reports) across
multiple files, the single/double mismatch triggered "Mismatch in type
of outputs" at index 2.

Fix: cast compVarPct and TESA rejPct to double at the two points where
they enter the report struct, so the type is always double regardless
of EEG.data precision.

+++ src/runPipeline.m
@@ -579,7 +579,8 @@ for nfile = 1:nFiles
                         data2D = reshape(EEG.data(EEG.icachansind,:,:), numel(EEG.icachansind), []);
                         totalVar = sum(var(data2D, 0, 2));
                         if totalVar > 0
-                            pendingICAStats.compVarPct = (var(act2D, 0, 2) / totalVar * 100)';
+                            % Cast to double: EEG.data may be single, making var() single.
+                            pendingICAStats.compVarPct = double((var(act2D, 0, 2) / totalVar * 100)');
                         end
                     end
                     if isfield(EEG,'etc') && isfield(EEG.etc,'ic_classification') && ...
@@ -978,7 +979,8 @@ for nfile = 1:nFiles
 
                     hasVars = isfield(cl, 'compVars') && numel(cl.compVars) >= numel(cl.compClass);
                     if hasVars && nRejTESA > 0
-                        rejPct         = cl.compVars(rejIdx);
+                        % Cast to double: TESA compVars may be single when EEG.data is single.
+                        rejPct         = double(cl.compVars(rejIdx));
                         rnd.varRemoved = sum(rejPct);
                         rnd.varMin     = min(rejPct);
                         rnd.varMax     = max(rejPct);
